+++ crates/core/action-pipeline/Cargo.toml
@@ -25,6 +25,7 @@ moon_utils = { path = "../utils" }
 moon_workspace = { path = "../workspace" }
 moonbase = { path = "../moonbase" }
 async-channel = "1.8.0"
+console = "0.15.2"
 rustc-hash = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
